Added debugger controller widgets to Xcode, and fixed minor compile warning.
git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@2939 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba

+++ src/debugger/gui/TiaOutputWidget.cxx
@@ -66,7 +66,7 @@ void TiaOutputWidget::loadConfig()
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 void TiaOutputWidget::saveSnapshot()
 {
-  int number = instance().getTicks() / 1000;
+  int number = (int)(instance().getTicks() / 1000);
   ostringstream sspath;
   sspath << instance().snapshotSaveDir()
          << instance().console().properties().get(Cartridge_Name)
